toks_inside is overwritten
Hi, 

thanks for this awesome package. Have been using it a lot in the past. 
I might be wrong here, but should it not be that we select the 10 word window around the keywords and then delete the keywords from these patterns, but not overwrite the entire tokens object with the full tokens object again? 
Proposed the relevant change above.